pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "techpdfparser"
version = "0.1.0"
description = "A technical PDF parsing library"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "pymupdf>=1.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "uv",
]
gpu = [
    "mineru>=2.5",
    "transformers>=4.40",
    "torch>=2.2",
    "beautifulsoup4>=4.12",
]

[tool.setuptools.packages.find]
include = ["extraction*"]
exclude = [
    "outputs*",
    "_archive*",
    "embedding*",
    "indexing*",
    "extraction.tests*",
    "tests*",
    "tasks*",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
files = [
    "extraction",
]
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["mineru.*", "transformers.*", "torch.*", "bs4.*", "fitz.*", "pymupdf.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["extraction/tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: heavy end-to-end tests that require MinerU / GPU / extra models",
]
